Adding rotator class - Added `Rotator()` class - Moved `promax()` and `varimax()` methods into this class. - Added `rotate()` method.
